버그 수정 및 패들에 대한 움직임 개선
기존에 어플이 먹통이던 버그를 수정
패들의 움직임을 터치로 구현
가속도를 이용한 공의 움직임에 변화를 주는 방식보다
공이 패들에 충돌하는 위치에 따라 움직임을 변화하는 방식 채택 후 구현

+++ app/src/main/java/com/example/breakingblock/Block.kt
@@ -15,7 +15,7 @@ class Block(
     }
 
     // 블럭에 충돌 확인(0: 충돌하지 않음, 1: 왼쪽, 2: 오른쪽, 3: 위, 4: 아래)
-    fun IsClash(Ball_X: Int, Ball_Y: Int, Ball_D: Int, Ball_R: Int): Int {
+    fun IsClash(Ball_X: Float, Ball_Y: Float, Ball_D: Int, Ball_R: Int): Int {
         // 블럭 좌측 충돌 확인
         if (Block_X - Ball_D <= Ball_X && Ball_X <= Block_X - Ball_R
             && Block_Y - Ball_R <= Ball_Y && Ball_Y <= Block_Y + Block_H - Ball_R
